refactor: 테스트를 위한 클래스인 CustomGenerator의 패키지 변경
